Commit contains implementation of Multi-columned layout, Criss-Cross animations,  and variable text size elements.
The Implementation of multi-columned layout is done by the use of a custom viewg
roup which just positions the elements in zig-zag order. The multi-sized height of the elements is maintained by the custom text view which adjusts its height based on the text. Initially, the size of the view is calculated based on a default text size, and if the height is less than the minimum text size, then the size of the text is increased until it reaches the height or the text size reaches the maximum value. If the height of the view is more than the maximum size, then the text size is decreased until the point where the height matches or the text size is reached to the minimum value. The criss-cross animations are done by the use of the layout transition object. The criss-cross effect is produced due to the different heights of the object in the layout and animating layout changes produces the criss-cross animation effect.